+++ packages/@o3r/chrome-devtools/src/app-devtools/app.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { NgbNavModule } from '@ng-bootstrap/ng-bootstrap';
+import { RulesetHistoryPresModule } from '@o3r/rules-engine';
 import { AppComponent } from './app.component';
 import { ConfigPanelPresModule } from './config-panel/config-panel-pres.module';
 import { DebugPanelPresModule } from './debug-panel/debug-panel-pres.module';
-import { RulesetHistoryPresModule } from '../components/rules-engine/ruleset-history/ruleset-history-pres.module';
 import { ComponentPanelPresModule } from './component-panel/component-panel-pres.module';
 import { AppConnectionModule } from '../components/app-connection/app-connection.module';
