+++ packages/@o3r/chrome-devtools/src/app-devtools/config-panel/config-panel-pres.component.ts
@@ -15,6 +15,12 @@ export class ConfigPanelPresComponent {
   public configs$: Observable<Dictionary<ConfigurationModel>>;
 
   constructor(connectionService: ChromeExtensionConnectionService) {
+    connectionService.sendMessage(
+      'requestMessages',
+      {
+        only: 'configurations'
+      }
+    );
     this.configs$ = connectionService.message$.pipe(
       filter(isConfigurationsMessage),
       map((message) => message.configurations)
