+++ packages/@o3r/chrome-devtools/src/components/otter-component/otter-component.module.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { NgbAccordionModule, NgbNavModule } from '@ng-bootstrap/ng-bootstrap';
 import { ConfigFormModule } from '../config-form/config-form.module';
-import { RulesetHistoryPresModule } from '../rules-engine/ruleset-history/ruleset-history-pres.module';
+import { RulesetHistoryPresModule } from '@o3r/rules-engine';
 import { ListPipe, NbPropPipe, OtterComponentComponent } from './otter-component.component';
