+++ packages/@o3r/rules-engine/src/components/index.ts
@@ -1,0 +1,1 @@
+export * from './rules-engine/index';
