+++ packages/@o3r/rules-engine/src/components/rules-engine/index.ts
@@ -1,0 +1,3 @@
+export * from './ruleset-history/ruleset-history-pres.module';
+export * from './ruleset-history/ruleset-history-pres.component';
+export * from './shared/index';
