+++ packages/@o3r/rules-engine/src/components/rules-engine/rule-actions/rule-actions-pres.style.scss
@@ -1,4 +1,4 @@
-@import '../shared/shared-mixin';
+@import '../styling/shared-mixin';
 
 o3r-rule-actions-pres {
   @include rule-debug-structure;
