+++ packages/@o3r/rules-engine/src/components/rules-engine/rule-condition/rule-condition-pres.template.html
