+++ packages/@o3r/rules-engine/src/components/rules-engine/rule-key-value/rule-key-value-pres.style.scss
@@ -1,4 +1,4 @@
-@import '../shared/shared-mixin';
+@import '../styling/shared-mixin';
 
 o3r-rule-key-value-pres {
   @include rule-debug-structure;
