+++ packages/@o3r/rules-engine/src/components/rules-engine/rule-tree/rule-tree-pres.component.ts
@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, Component, Input, ViewEncapsulation } from '@angular/core';
-import type { AllBlock, TopLevelCondition } from '@o3r/rules-engine';
+import type { AllBlock, TopLevelCondition } from '../../../engine';
 
 @Component({
   selector: 'o3r-rule-tree-pres',
