+++ packages/@o3r/rules-engine/src/components/rules-engine/rule-tree/rule-tree-pres.style.scss
@@ -1,4 +1,4 @@
-@import '../shared/shared-mixin.scss';
+@import '../styling/shared-mixin';
 
 $border-action-panel: 1px solid #999999;
 $tree-line: 1px dashed;
