+++ packages/@o3r/rules-engine/src/components/rules-engine/ruleset-history/ruleset-history-pres.module.ts
@@ -4,7 +4,7 @@ import { RuleActionsPresComponent } from '../rule-actions/rule-actions-pres.comp
 import { RuleConditionPresComponent } from '../rule-condition/rule-condition-pres.component';
 import { RuleKeyValuePresComponent } from '../rule-key-value/rule-key-value-pres.component';
 import { RuleTreePresComponent } from '../rule-tree/rule-tree-pres.component';
-import { FallbackToPipe } from '../shared/fallback-to.pipe';
+import { FallbackToPipe } from '../shared';
 import { RulesetHistoryPresComponent } from './ruleset-history-pres.component';
 
 @NgModule({
