+++ packages/@o3r/rules-engine/src/components/rules-engine/ruleset-history/ruleset-history-pres.style.scss
@@ -1,4 +1,4 @@
-@import '../shared/shared-mixin';
+@import '../styling/shared-mixin';
 
 o3r-ruleset-history-pres {
   @include rule-debug-structure;
