+++ packages/@o3r/rules-engine/src/components/rules-engine/styling/shared-mixin.scss
