+++ packages/@o3r/rules-engine/src/public_api.ts
@@ -1,3 +1,4 @@
+export * from './components/index';
 export * from './devkit/index';
 export * from './engine/index';
 export * from './fact/index';
