+++ tsconfig.build.json
@@ -28,7 +28,6 @@
       "@o3r/vscode-extension": ["packages/@o3r/vscode-extension/dist", "packages/@o3r/vscode-extension/src/public_api"],
       "@o3r/storybook": ["packages/@o3r/storybook/dist", "packages/@o3r/storybook/src/public_api"],
       "@o3r/third-party": ["packages/@o3r/third-party/dist", "packages/@o3r/third-party/src/public_api"],
-      "@o3r/chrome-devtools": ["packages/@o3r/chrome-devtools/dist", "packages/@o3r/chrome-devtools/src/public_api"],
       "@o3r/apis-manager": ["packages/@o3r/apis-manager/dist", "packages/@o3r/apis-manager/src/public_api"],
       "@o3r/application": ["packages/@o3r/application/dist", "packages/@o3r/application/src/public_api"],
       "@o3r/mobile": ["packages/@o3r/mobile/dist", "packages/@o3r/mobile/src/public_api"],
